% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm-package.R
\docType{package}
\name{gsm.qc-package}
\alias{gsm.qc}
\alias{gsm.qc-package}
\title{gsm.qc: Good Statistical Monitoring Testing Suite}
\description{
The Good Statistical Monitoring or 'gsm' R packages provides a framework for statistical data monitoring. This package provides all a testing suite for all the packages involved.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Gilead-BioStats/gsm.qc}
  \item \url{https://gilead-biostats.github.io/gsm.qc}
  \item Report bugs at \url{https://github.com/Gilead-BioStats/gsm.qc/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Wildfire \email{jwildfire@gmail.com}

Authors:
\itemize{
  \item Zelos Zhu \email{zelos.zhu@atorusresearch.com}
}

Other contributors:
\itemize{
  \item Gilead Sciences [copyright holder]
}

}
\keyword{internal}
