% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.qc-package.R
\docType{package}
\name{gsm.qc-package}
\alias{gsm.qc}
\alias{gsm.qc-package}
\title{gsm.qc: Good Statistical Monitoring Testing Suite}
\description{
The Good Statistical Monitoring or 'gsm' R packages provides a framework for statistical data monitoring. This package provides a testing suite for all the packages involved and is the source of the qualification testing report.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Gilead-BioStats/gsm.qc}
  \item \url{https://gilead-biostats.github.io/gsm.qc}
  \item Report bugs at \url{https://github.com/Gilead-BioStats/gsm.qc/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Wildfire \email{jwildfire@gmail.com}

Authors:
\itemize{
  \item Zelos Zhu \email{zelos.zhu@atorusresearch.com}
  \item Laura Maxwell \email{lkmaxwell23@gmail.com}
}

Other contributors:
\itemize{
  \item Gilead Sciences [copyright holder]
}

}
\keyword{internal}
